# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact:
    name: Ethan Swan
  title: Identity Service
  version: "1.0"
paths:
...
  /login:
    get:
      description: Displays the login form with OAuth parameters preserved in hidden
        fields
      parameters:
      - description: Registered client identifier
        in: query
        name: client_id
        required: true
        type: string
      - description: Where to send the user after auth
        in: query
        name: redirect_uri
        required: true
        type: string
      - description: CSRF protection token
        in: query
        name: state
        required: true
        type: string
      - description: Requested scopes (e.g., 'openid profile email')
        in: query
        name: scope
        required: true
        type: string
      - description: PKCE code challenge (SHA256 hash)
        in: query
        name: code_challenge
        required: true
        type: string
      - description: PKCE challenge method (must be 'S256')
        in: query
        name: code_challenge_method
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML login page
          schema:
            type: string
      summary: Show login page
      tags:
      - authentication
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Validates username/password, creates authenticated session, generates
        authorization code, and redirects to client
      parameters:
      - description: User username
        in: formData
        name: username
        required: true
        type: string
      - description: User password
        in: formData
        name: password
        required: true
        type: string
      - description: OAuth client ID
        in: formData
        name: client_id
        required: true
        type: string
      - description: OAuth redirect URI
        in: formData
        name: redirect_uri
        required: true
        type: string
      - description: CSRF protection token
        in: formData
        name: state
        required: true
        type: string
      - description: Requested scopes
        in: formData
        name: scope
        required: true
        type: string
      - description: PKCE code challenge
        in: formData
        name: code_challenge
        required: true
        type: string
      - description: PKCE challenge method
        in: formData
        name: code_challenge_method
        required: true
        type: string
      produces:
      - text/html
      responses:
        "302":
          description: Redirect to redirect_uri with authorization code
          schema:
            type: string
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
      summary: Process login
      tags:
      - authentication
...
  /register:
    get:
      description: Displays the registration form for creating a new user account.
        Accepts OAuth parameters to preserve them through the registration flow.
      parameters:
      - description: OAuth client identifier (preserved through flow)
        in: query
        name: client_id
        type: string
      - description: OAuth redirect URI (preserved through flow)
        in: query
        name: redirect_uri
        type: string
      - description: OAuth state parameter (preserved through flow)
        in: query
        name: state
        type: string
      - description: OAuth scope (preserved through flow)
        in: query
        name: scope
        type: string
      - description: PKCE code challenge (preserved through flow)
        in: query
        name: code_challenge
        type: string
      - description: PKCE challenge method (preserved through flow)
        in: query
        name: code_challenge_method
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML registration page
          schema:
            type: string
      summary: Show registration page
      tags:
      - authentication
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Creates a new user account with username, email, and password
      parameters:
      - description: Desired username
        in: formData
        name: username
        required: true
        type: string
      - description: User email address
        in: formData
        name: email
        required: true
        type: string
      - description: User password (min 8 characters)
        in: formData
        name: password
        required: true
        type: string
      - description: Password confirmation
        in: formData
        name: confirm_password
        required: true
        type: string
      produces:
      - text/html
      responses:
        "302":
          description: Redirect to login page on success
          schema:
            type: string
        "400":
          description: Invalid request parameters or validation error
          schema:
            type: string
        "409":
          description: Username or email already exists
          schema:
            type: string
      summary: Register new user
      tags:
      - authentication
...
swagger: "2.0"